Moved diff writers into their own file, cleaned up a little
